requirements.txt
click==6.6
Flask==0.11.1
httpie==0.9.6
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pkg-resources==0.0.0
Pygments==2.1.3
pyserial==3.2.1
requests==2.11.1
Werkzeug==0.11.11
